% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{assignCluster}
\alias{assignCluster}
\title{Make Cluster Assignments}
\usage{
assignCluster(X, centers)
}
\arguments{
\item{X}{matrix of observations}

\item{centers}{matrix of cluster centers}
}
\value{
A vector of cluster labels
}
\description{
Assigns observation to the nearest cluster center, using squared Euclidean distance.
}
\author{
Joshua Keller
}

