% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{createPCAmodelmatrix}
\alias{createPCAmodelmatrix}
\title{Get PCA comps for regression}
\usage{
createPCAmodelmatrix(data, ncomps, covarnames = colnames(data),
  center = TRUE, scale = TRUE, matrixonly = TRUE)
}
\arguments{
\item{data}{Matrix or data frame of data}

\item{ncomps}{Number of PCA components to return.}

\item{covarnames}{Names of variables or column numbers in \code{data}
    on which the PCA is to be run.}

\item{center}{Logical indicator of whether \code{data} should be centered. Passed to \code{\link{prcomp}}.}

\item{scale}{Logical indicator of whether \code{data} should be scaled. Passed to \code{\link{prcomp}}.}

\item{matrixonly}{Logical indicator of whether only the model matrix should
be returned, or the full output from \code{\link{prcomp}}.}
}
\description{
Wrapper function for creating PCA components to be used
		in a regression analysis.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{createTPRSmodelmatrix}}
}

