% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{createTPRSmodelmatrix}
\alias{createTPRSmodelmatrix}
\title{Get matrix of TPRS for regression}
\usage{
createTPRSmodelmatrix(data, df = 5, covarnames = NULL, xname = "x",
  yname = "y", TPRSfx = TRUE, matrixonly = TRUE)
}
\arguments{
\item{data}{Matrix or data frame of data}

\item{df}{Degrees of freedom for thinplate splines. This does not include an intercept, so the \code{k} argument of \code{s()} is \eqn{k = df + 1}.}

\item{covarnames}{Names of other covariates to be included in the model
    matrix.}

\item{xname}{Name of variable the provides the x-coordinate of location.}

\item{yname}{Name of variable the provides the y-coordinate of location.}

\item{TPRSfx}{Should the TPRS degrees of freedom be fixed. Passed as the \code{fx} argument to \code{s()}.}

\item{matrixonly}{Logical indicator of whether only the model matrix should
be returned, or the full output from \code{\link{gam}}.}
}
\description{
Wrapper function for creating thin-plate regression splines (TPRS)
		to be used in a regression analysis.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{createPCAmodelmatrix}}
}

