% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{fitCV}
\alias{fitCV}
\title{Cross-validation of Predictive K-means Clustering}
\usage{
fitCV(X, R, K, cv.groups = 10, doKmeans = FALSE, sigma2 = 0,
  sigma2fixed = FALSE, scale = TRUE, covarnames = NULL, PCA = FALSE,
  pcacovarnames = colnames(R), npca = 5, TPRS = TRUE, TPRSdf = 5,
  TPRSxname = "x", TPRSyname = "y", mu = NULL, muStart = c("kmeans",
  "random"), maxitEM = 100, tol = 1e-05, convEM = c("both", "mu",
  "gamma"), nStarts = 1, maxitMlogit = 500, verbose = 0,
  muRestart = 1000, returnAll = FALSE, ...)
}
\arguments{
\item{X}{Outcome data}

\item{R}{Covariates}

\item{K}{Number of clusters}

\item{cv.groups}{A list or matrix providing the cross-validation
    groups for splitting the data.  Alternatively, a single
    number giving the number of groups into which the
    data are randomly split. A value of '0' implies leave-one-out.
    Defaults to 10.}

\item{doKmeans}{Should kmeans also be done, in addition to predictive
    kmeans? Note: If sigma2=0 and sigma2fixed=TRUE, then doKmeans is
    is set to TRUE, and no predictive k-means is done.}

\item{sigma2}{starting value of sigma2 (ADD SPECIAL CASE OF 0 here)}

\item{sigma2fixed}{Logical indicating whether sigma2
should be held fixed.  If FALSE, then
sigma2 is estimated using Maximum Likelihood.}

\item{scale}{Should the outcomes be re-scaled within each training
group?}

\item{mu}{= starting values for centers. If NULL (default),
then value is chosen according to \code{muStart}.}

\item{muStart}{Character string indicating how initial value
 of mu should be selected. Only used if mu=NULL.  Possible
values are "random" (default) or "kmeans".}

\item{maxitEM}{Maximum number of EM iterations for
finding the Mixture of Experts solution}

\item{tol}{convergence criterion}

\item{convEM}{controls the measure of convergence for the
EM algorithm.  Should be one of "mu", "gamma", or
    "both".  Defaults to "both."  The EM algorithm
stops when the Frobenius norm of the change in mu,
    the change in gamma, or the change in mu and the change in gamma
    is less than 'tol'.}

\item{nStarts}{number of times to perform EM algorithm}

\item{maxitMlogit}{Maximum number of iterations in the=
mlogit optimization (nested within EM algorithm)}

\item{verbose}{numeric vector indicating how much output to produce}

\item{muRestart}{Gives max number of attempts at picking
   starting values. Only used when muStart='random'.
    If selected starting values for mu are constant
within each cluster, then the starting values
are re-selected up to muRestart times.}

\item{returnAll}{A list containing all \code{nStarts} solutions is
included in the output.}

\item{...}{Additional arguments passed to \code{\link{mlogit}}}
}
\description{
Performs cross-validation of predictive-kmeans on a dataset.
}
\details{
To be added....
}
\author{
Joshua Keller
}
\seealso{
\code{\link{predkmeans}}, \code{\link{predictCV}}

Other 'predkmeans methods': \code{\link{predkmeans}}
}

