% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{predictCV}
\alias{predictCV}
\title{Prediction from Cross-validation of Predictive K-means Clustering}
\usage{
predictCV(fitCV, X, R, method = c("ML", "MixExp", "SVM"), doMetrics = FALSE,
  covarnames = NULL, PCA = FALSE, pcacovarnames = colnames(R), npca = 5,
  TPRS = TRUE, TPRSdf = 5, TPRSxname = "x", TPRSyname = "y", ...)
}
\description{
Performs cross-validation of predictive-kmeans on a dataset.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{fitCV}}
}

