% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_predict.R
\name{predictionMetrics}
\alias{predictionMetrics}
\title{Measures of Prediction Performance}
\usage{
predictionMetrics(centers, cluster.pred, X)
}
\arguments{
\item{centers}{Matrix of Cluster centers}

\item{cluster.pred}{vector of predicted cluster membership.  Assumed to be integers
                corresponding to rows of \code{centers}.}

\item{X}{Matrix of observations at prediction locations.}
}
\value{
A list with the following elements:
\item{cluster.pred}{Predicted cluster assignments}
\item{cluster.assign}{'Best' cluster assignments (i.e. assignment to closest cluster center)}
\item{pred.acc}{Proportion of cluster labels correctly predicted.}
\item{MSEP}{Mean squared error of prediction.  Sum of squared distances between predicted cluster center and best (i.e. closest) cluster center.}
\item{wSS}{Within-cluster sum-of-squares.  Sum of squared distances between observations at prediction locations and best (i.e. closest) cluster center.}
\item{GPE}{Global Prediction Error (term needs re-naming). Sum of squared distances between observation at prediciton locations and predicted cluster center.}
}
\description{
Computes several measures of performance for cluster label prediction.
}
\seealso{
\code{\link{predictML}}
}

