% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_predict.R
\name{predictML.predkmeans}
\alias{predictML}
\alias{predictML.predkmeans}
\alias{predictSVM}
\alias{predictSVM.predkmeans}
\title{Prediction of Cluster Membership}
\usage{
predictML.predkmeans(object = NULL, centers = object$centers,
  K = nrow(centers), R, Rstar, X = NULL, Xstar = NULL,
  tr.assign = object$cluster, muStart = "random", maxitMlogit = 500,
  verbose = 1, nMlogitStarts = 1,
  mlogit.control = list(suppressFittedWarning = TRUE))

predictSVM.predkmeans(object = NULL, centers = object$centers, R, Rstar,
  K = nrow(centers), X = NULL, Xstar = NULL, tr.assign = object$cluster,
  svm.control = list(gamma = c(1/(2:1), 2), cost = seq(20, 100, by = 20)))
}
\arguments{
\item{object}{A predkmeans object, from which the cluster centers will be extracted.}

\item{centers}{Matrix of cluster centers, assumed to be K-by-p}

\item{K}{Number of clusters}

\item{R}{matrix of covariates for observations to be predicted at.}

\item{Rstar}{matrix of covariates at training locations}

\item{X}{Optional matrix of observations at prediction locations. If provided, measures of prediction performance will be reported.}

\item{Xstar}{matrix of observation at training locations.  Either this or \code{tr.assign} is required.}

\item{tr.assign}{vector of cluster assignments at training locations}

\item{muStart}{starting value for cluster centers in mlogit optimization (IDEA: change to pull from predkmeans object?).  If not provided, starting values are selected randomly.}

\item{maxitMlogit}{Maximum number of iterations for \code{mlogit} in prediction}

\item{verbose}{integer indicating amount of output to be displayed}

\item{nMlogitStarts}{number of mlogit starts to use in estimation of parameters}

\item{mlogit.control}{list of control parameters to be passes to \code{mlogit}}

\item{svm.control}{list of options for \code{best.svm}}
}
\value{
A list containing some or all of the following elements:
	\item{tr.assign}{Cluster assignments at training locations}
	\item{mlfit}{A subset of the mlogit object returned by the function of that name}
\item{beta}{beta}
\item{test.pred}{test.pred}
\item{test.assign}{test.assign}
\item{pred.acc}{pred.acc}
\item{MSEP}{MSEP}
\item{wSS.subj}{wSS.subj}
}
\description{
Predicts cluster membership using either multinomial logistic
   regression or SVMs.
}
\details{
Function for predicting cluster membership in clusters identified by k-means or predictive k-means using multinomial logistic regression or support vector machines (SVMs).  For multinomial logitic regression, parameter estimation is handled by \code{mlogit}.  For SVMs, the svm is fit using \code{best.svm} in \code{e1071} package.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{mlogit}}, \code{\link{predkmeans}}
}

