% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predkmeans.R
\docType{package}
\name{predkmeans-package}
\alias{predkmeans-package}
\title{Covariate Adaptive Clustering}
\description{
Clusters multivariate exposures, using a mixture of experts model to allow covariates to influence cluster centers. Motivated by air pollution epidemiology settings, where cluster membership needs to be predicted across space. Includes functions for predicting cluster membership using spatial splines and principal component analysis (PCA) scores using either multinomial logistic regression or support vector machines (SVMs). For method details see Keller et al. (2017) <doi:10.1214/16-AOAS992>
}
