% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{predkmeansCVest}
\alias{predkmeansCVest}
\title{Cross-validation of Predictive K-means Clustering}
\usage{
predkmeansCVest(X, R, K, cv.groups = 10, sigma2 = 0, sigma2fixed = FALSE,
  scale = TRUE, covarnames = NULL, PCA = FALSE,
  PCAcontrol = list(covarnames = colnames(R), ncomps = 5), TPRS = TRUE,
  TPRScontrol = list(df = 5, xname = "x", yname = "y"), returnAll = FALSE,
  ...)
}
\arguments{
\item{X}{Outcome data}

\item{R}{Covariates. Coerced to data frame.}

\item{K}{Number of clusters}

\item{cv.groups}{A list or matrix providing the cross-validation
groups for splitting the data.  Alternatively, a single
number giving the number of groups into which the
data are randomly split. A value of '0' implies leave-one-out.
Defaults to 10.}

\item{sigma2}{starting value of sigma2. Note: If sigma2=0 and 
sigma2fixed=TRUE, then regular k-means is done in place of  
predictive k-means.}

\item{sigma2fixed}{Logical indicating whether sigma2
should be held fixed.  If FALSE, then
sigma2 is estimated using Maximum Likelihood.}

\item{scale}{Should the outcomes be re-scaled within each training
group?}

\item{covarnames}{Names of covariates to be included directly.}

\item{PCA}{Logical indicator for whether PCA components should be computed
from R.}

\item{PCAcontrol}{Arguments passed to \code{\link{createPCAmodelmatrix}}. This includes \code{ncomps}.}

\item{TPRS}{Logical indicator for whether thin-plate regression
splines should be created and added to covariates.}

\item{TPRScontrol}{Arguments passed to \code{\link{createTPRSmodelmatrix}}. This includes \code{df}.}

\item{returnAll}{A list containing all \code{nStarts} solutions is
included in the output.}

\item{...}{Additional arguments passed to \code{\link{predkmeans}}}
}
\description{
Performs cross-validation of predictive-kmeans on a dataset.
}
\details{
To be added....
}
\author{
Joshua Keller
}
\seealso{
\code{\link{predkmeans}}, \code{\link{predkmeansCVpred}}, \code{\link{createPCAmodelmatrix}}, \code{\link{createTPRSmodelmatrix}}

Other 'predkmeans methods': \code{\link{predkmeans}},
  \code{\link{print.predkmeans}},
  \code{\link{relevel.predkmeans}}
}

