% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{predkmeansCVpred}
\alias{predkmeansCVpred}
\title{Prediction from Cross-validation of Predictive K-means Clustering}
\usage{
predkmeansCVpred(object, X = object$X, R = object$R, method = c("ML",
  "MixExp", "SVM"), ...)
}
\arguments{
\item{object}{A \code{predkmeansCVest} object. See \code{\link{predkmeansCVest}}.}

\item{X}{Matrix of observations}

\item{R}{matrix of covariates.}

\item{method}{Character string indicating which prediciton method should be used. Optins are \code{ML}, \code{MixExp}, and \code{SVM}. See \code{\link{predictML}} for more information.}

\item{...}{Additional arguments passed to the prediction method.}
}
\description{
Performs cross-validation of predictive-kmeans on a dataset.
}
\seealso{
\code{\link{predkmeansCVest}}, \code{\link{predictML}}
}
\author{
Joshua Keller
}
