% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_mlogit.R
\name{print.mlogit}
\alias{print.mlogit}
\title{Print details for class \code{mlogit}}
\usage{
\method{print}{mlogit}(x, ...)
}
\arguments{
\item{x}{object of class \code{mlogit}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{print}} method for class \code{mlogit}.
}
\seealso{
Other mlogit methods: \code{\link{mlogit}};
  \code{\link{summary.mlogit}}
}

