% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{print.predkmeansCVpred}
\alias{print.predkmeansCVpred}
\title{Print details for class \code{predkmeansCVpred}}
\usage{
\method{print}{predkmeansCVpred}(x, ...)
}
\arguments{
\item{x}{object of class \code{predkmeansCVpred}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{print}} method for class \code{predkmeansCVpred}.
}

