% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_predkmeans.R
\name{relevel.predkmeans}
\alias{relevel.predkmeans}
\title{Re-order cluster labels}
\usage{
\method{relevel}{predkmeans}(x, ref = NULL, order = NULL, ...)
}
\arguments{
\item{x}{object of class \code{predkmeans}}

\item{ref}{New reference group ("Cluster 1"). Only used if \code{order} is NULL.}

\item{order}{New order of clusters.}

\item{...}{Ignored additional arguments.}
}
\description{
Function for re-ordering the order of clusters in a predkmeans object.
}
\details{
The elements of the \code{order} argument should refer
	to the current position of clusters, with the position
	giving the new order. So 'c(3, 1, 2)' moves 1 to 2, 2 to 3, and 3 to 1.
}
\seealso{
Other methods for predkmeans objects: \code{\link{predictML.predkmeans}},
  \code{\link{print.predkmeans}},
  \code{\link{summary.predkmeans}}
}
\author{
Joshua Keller
}
