% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_mlogit.R
\name{summary.mlogit}
\alias{summary.mlogit}
\title{Compute summary details for class \code{mlogit}}
\usage{
\method{summary}{mlogit}(object, ...)
}
\arguments{
\item{object}{object of class \code{mlogit}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{summary}} method for class \code{mlogit}.
}
\seealso{
Other mlogit methods: \code{\link{mlogit}},
  \code{\link{print.mlogit}}
}

