% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cv.R
\name{summary.predkmeansCVpred}
\alias{summary.predkmeansCVpred}
\title{Compute summary details for class \code{predkmeansCVpred}}
\usage{
\method{summary}{predkmeansCVpred}(object, ...)
}
\arguments{
\item{object}{object of class \code{predkmeansCVpred}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{summary}} method for class \code{predkmeansCVpred}.
}

