% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_predict.R
\name{predictionMetrics}
\alias{predictionMetrics}
\title{Measures of Prediction Performance}
\usage{
predictionMetrics(centers, cluster.pred, X, labels = TRUE)
}
\arguments{
\item{centers}{Matrix of Cluster centers}

\item{cluster.pred}{Vector of predicted cluster membership. Should be integers
or names corresponding to rows of \code{centers}.}

\item{X}{Matrix of observations at prediction locations.}

\item{labels}{Logical indicating whether cluster prediction and}
}
\value{
A list with the following elements:
\item{cluster.pred}{Predicted cluster assignments (same as argument provided).}
\item{cluster.assign}{Integer vector of 'best' cluster assignments (i.e. assignment to closest cluster center)}
\item{MSPE}{Mean squared prediction error. Sum of squared distances between observations and predicted cluster centers.}
\item{wSS}{Within-cluster sum-of-squares.  Sum of squared distances between observations at prediction locations and best (i.e. closest) cluster center.}
\item{pred.acc}{Proportion of cluster labels correctly predicted.}
\item{MSME}{Mean squared misclassification error.  Sum of squared distances between predicted cluster center and best (i.e. closest) cluster center.}
}
\description{
Computes several measures of performance for cluster label prediction.
}
\seealso{
\code{\link{predictML}}
}
\author{
Joshua Keller
}
