% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_predict.R
\name{predictML.predkmeans}
\alias{predictML}
\alias{predictML.predkmeans}
\alias{predictMixExp}
\alias{predictMixExp.predkmeans}
\alias{predictSVM}
\alias{predictSVM.predkmeans}
\title{Prediction of Cluster Membership}
\usage{
\method{predictML}{predkmeans}(object = NULL, centers = object$centers,
  K = nrow(centers), R, Rstar, Xstar = NULL, tr.assign = object$cluster,
  muStart = "random", maxitMlogit = 500, verbose = 1, nMlogitStarts = 1,
  mlogit.control = list(suppressFittedWarning = TRUE))

\method{predictSVM}{predkmeans}(object = NULL, centers = object$centers, R,
  Rstar, K = nrow(centers), Xstar = NULL, tr.assign = object$cluster,
  svm.control = list(gamma = c(1/(2:1), 2), cost = seq(20, 100, by = 20)))

\method{predictMixExp}{predkmeans}(object, R, Rstar = NULL)
}
\arguments{
\item{object}{A predkmeans object, from which the cluster centers will be extracted.}

\item{centers}{Matrix of cluster centers, assumed to be K-by-p}

\item{K}{Number of clusters}

\item{R}{matrix of covariates for observations to be predicted at.}

\item{Rstar}{matrix of covariates at training locations}

\item{Xstar}{matrix of observation at training locations.  Either this or \code{tr.assign} is required.}

\item{tr.assign}{vector of cluster assignments at training locations.  By default, extracted from \code{object}.}

\item{muStart}{starting value for cluster centers in mlogit optimization (IDEA: change to pull from predkmeans object?).  If not provided, starting values are selected randomly.}

\item{maxitMlogit}{Maximum number of iterations for \code{mlogit} in prediction}

\item{verbose}{integer indicating amount of output to be displayed}

\item{nMlogitStarts}{number of mlogit starts to use in estimation of parameters}

\item{mlogit.control}{list of control parameters to be passes to \code{mlogit}}

\item{svm.control}{list of options for \code{best.svm}}
}
\value{
A list containing some or all of the following elements:
\item{tr.assign}{Cluster assignments at training locations}
	\item{mlfit}{A subset of the mlogit object returned by the function of that name}
\item{beta}{Estimated model parameters}
\item{test.pred}{Predicted cluster assignments at test locations}
}
\description{
Predicts cluster membership using either multinomial logistic
   regression or SVMs.
}
\details{
Function for predicting cluster membership in clusters identified by k-means or predictive k-means using multinomial logistic regression or support vector machines (SVMs).  For multinomial logitic regression, parameter estimation is handled by \code{mlogit}.  For SVMs, the svm is fit using \code{best.svm} in \code{e1071} package.

Because this prediction includes return information about cluster assignment
and prediction model parameters, this method is deliberately distinct from
the generic \code{predict} functions.

The \code{predictMixExp} funciton provides predictions from 
the 'working' cluster assignments created as part of the 
mixture of experts algorithm from \code{predkmeans}.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{mlogit}}, \code{\link{predkmeans}}, \code{\link{predictionMetrics}}
}

