% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_predkmeans.R
\name{summary.predkmeans}
\alias{summary.predkmeans}
\title{Compute summary details for class \code{predkmeans}}
\usage{
\method{summary}{predkmeans}(object, ...)
}
\arguments{
\item{object}{object of class \code{predkmeans}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{summary}} method for class \code{predkmeans}.
}
\seealso{
Other predkmeans methods: \code{\link{summary.predkmeansCVpred}}
}

